refactor lazy structure
